requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
bleach==3.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
orbit-predictor==1.12.0
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pyrtlsdr==0.2.92
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
regex==2020.2.20
requests==2.23.0
schedule==0.6.0
scipy==1.4.1
Send2Trash==1.5.0
sgp4==2.5
six==1.14.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
zipp==3.1.0
